requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.1
asgiref==3.11.1
attrs==25.3.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.0
cloudinary==1.44.0
colorama==0.4.6
cryptography==44.0.3
distlib==0.3.9
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emoji==2.14.1
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httptools==0.7.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.1
psycopg2==2.9.11
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic>=2.7.0
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.2
pytesseract==0.3.13
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
pyttsx==1.1
PyYAML==6.0.2
qrcode==8.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.5
starlette==0.46.2
twilio==9.6.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
watchfiles==1.1.1
websockets==15.0.1
yarl==1.20.0

gunicorn
whitenoise[brotli]
dj-database-url
